+++ verify_v2/src/vonage_verify_v2/__init__.py
@@ -8,13 +8,14 @@ from .requests import (
     VoiceChannel,
     WhatsappChannel,
 )
-from .responses import StartVerificationResponse
+from .responses import CheckCodeResponse, StartVerificationResponse
 from .verify_v2 import VerifyV2
 
 __all__ = [
     'VerifyV2',
     'VerifyError',
     'ChannelType',
+    'CheckCodeResponse',
     'Locale',
     'VerifyRequest',
     'SilentAuthChannel',